Rename SDL frontend. Add Stub Executable.
