+++ .travis/linux/upload.sh
@@ -9,6 +9,7 @@ COMPRESSION_FLAGS="-cJvf"
 mkdir "$REV_NAME"
 
 cp build/src/citra/citra "$REV_NAME"
+cp build/src/citra_cmd/citra-cmd "$REV_NAME"
 cp build/src/dedicated_room/citra-room "$REV_NAME"
 cp build/src/citra_qt/citra-qt "$REV_NAME"
