+++ src/CMakeLists.txt
@@ -10,6 +10,7 @@ add_subdirectory(input_common)
 add_subdirectory(tests)
 if (ENABLE_SDL2)
     add_subdirectory(citra)
+    add_subdirectory(citra_cmd)
 endif()
 if (ENABLE_QT)
     add_subdirectory(citra_qt)
