+++ src/citra_cmd/config.cpp
@@ -6,8 +6,8 @@
 #include <unordered_map>
 #include <SDL.h>
 #include <inih/cpp/INIReader.h>
-#include "citra/config.h"
-#include "citra/default_ini.h"
+#include "citra_cmd/config.h"
+#include "citra_cmd/default_ini.h"
 #include "common/file_util.h"
 #include "common/logging/log.h"
 #include "common/param_package.h"
