+++ src/citra_cmd/emu_window/emu_window_sdl2.cpp
@@ -8,7 +8,7 @@
 #define SDL_MAIN_HANDLED
 #include <SDL.h>
 #include <glad/glad.h>
-#include "citra/emu_window/emu_window_sdl2.h"
+#include "citra_cmd/emu_window/emu_window_sdl2.h"
 #include "common/logging/log.h"
 #include "common/scm_rev.h"
 #include "common/string_util.h"
