+++ src/citra_cmd/resource.h
